Created FeedImagesMapper class as helper for RemoteFeedLoader
